Partially revert pawns storm bug fix
Try to save space and use the minimum size
possible.

In particular restore int16_t for values and int8_t
for halfOpenFiles.

Use int16_t for storm values insted of int and also
instead of original buggy and too small int8_t.

+++ src/pawns.h
@@ -58,9 +58,9 @@ private:
 
   Key key;
   Bitboard passedPawns;
-  int mgValue, egValue;
-  int ksStormValue[2], qsStormValue[2];
-  int halfOpenFiles[2];
+  int16_t mgValue, egValue;
+  int16_t ksStormValue[2], qsStormValue[2];
+  uint8_t halfOpenFiles[2];
 };
